# public/.well-known/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sunu Yoon - API Covoiturage Sénégal",
    "description": "API pour rechercher, publier et réserver des covoiturages au Sénégal. Couvre toutes les villes: Dakar, Thiès, Saint-Louis, Touba, Kaolack, Ziguinchor, Mbour, Saly, etc. Prix en FCFA (XOF). Paiement Orange Money, Wave ou espèces.",
    "version": "1.0.0",
    "contact": {
      "name": "Sunu Yoon",
      "email": "support@sunuyoon.sn",
      "url": "https://sunuyoon.net"
    }
  },
  "servers": [
    {
      "url": "https://sunuyoon.net/api",
      "description": "Production"
    }
  ],
  "paths": {
    "/rides/recent": {
      "get": {
        "operationId": "getRecentRides",
        "summary": "Obtenir les trajets récents disponibles au Sénégal",
        "description": "Retourne les derniers trajets publiés avec places disponibles. Chaque trajet inclut: ville de départ, destination, date/heure, prix en FCFA, nombre de places, modèle de voiture, et informations du conducteur.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 10 },
            "description": "Nombre maximum de trajets à retourner"
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des trajets récents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "rides": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/Ride" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/guest-bookings": {
      "post": {
        "operationId": "createGuestBooking",
        "summary": "Réserver un trajet sans compte",
        "description": "Permet à un passager de réserver des places sur un trajet sans créer de compte. Nécessite nom, téléphone et méthode de paiement.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["rideId", "seats", "passengerName", "passengerPhone", "paymentMethod"],
                "properties": {
                  "rideId": { "type": "string", "description": "ID du trajet à réserver" },
                  "seats": { "type": "integer", "minimum": 1, "description": "Nombre de places" },
                  "passengerName": { "type": "string", "description": "Nom du passager" },
                  "passengerPhone": { "type": "string", "description": "Téléphone du passager (format +221XXXXXXXXX)" },
                  "paymentMethod": {
                    "type": "string",
                    "enum": ["WAVE", "ORANGE_MONEY", "CASH"],
                    "description": "Méthode de paiement: Wave, Orange Money, ou espèces"
                  },
                  "notes": { "type": "string", "description": "Notes optionnelles pour le conducteur" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Réservation créée avec succès"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Ride": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "driver": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "name": { "type": "string" },
              "rating": { "type": "number" },
              "reviewCount": { "type": "integer" },
              "isVerified": { "type": "boolean" }
            }
          },
          "origin": { "type": "string", "description": "Ville de départ" },
          "destination": { "type": "string", "description": "Ville d'arrivée" },
          "departureTime": { "type": "string", "format": "date-time" },
          "price": { "type": "number", "description": "Prix par place en FCFA (XOF)" },
          "currency": { "type": "string", "enum": ["XOF"] },
          "seatsAvailable": { "type": "integer" },
          "totalSeats": { "type": "integer" },
          "carModel": { "type": "string" },
          "description": { "type": "string" },
          "features": { "type": "array", "items": { "type": "string" } },
          "estimatedDuration": { "type": "string" }
        }
      },
...
    }
  }
}